Extract MultiAvatarManager to its own class
