+++ Editor/MultiAvatarManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31b57f7c4e1f87f4dbe09da94e788cea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
